fix(ci): python-coverage-shards — split init_ops.py under maintainability line limit
The forge-detection work (#539/#541) grew src/awf/cli/init_ops.py to 1815
lines, tripping test_first_party_code_files_stay_under_line_limit (1500-line
cap). Extract the self-contained local-service initialization helpers into a
new awf.cli.service_init_ops module, leaving project onboarding + forge
detection in init_ops. init_ops re-exports the moved names (combined import +
__all__) so every existing importer and test keeps referencing them via
awf.cli.init_ops unchanged. Pure move: both files now fall under the limit
(882 / 1061 lines).
